message

+++ packageKPS.py
@@ -109,12 +109,12 @@ def get_data_from_json(kdata):
     kpdf['D28'] = kdata['report.turnover_other']
 
     #% μεταβολής κύκλου εργασιών ανά δραστηριότητα ανά έτος
-    kpdf['D29'] = round((kdata['report.turnover_total'].astype(int).pct_change()*100),1)
+    kpdf['D29'] = round((kdata['report.turnover_total'].astype(float).pct_change()*100),1)
 
-    kpdf['D30'] = round((kpdf['D26'].astype(int).pct_change()*100),1)
-    kpdf['D31'] = round((kpdf['D27'].astype(int).pct_change()*100),1)
-    kpdf['D32'] = round((kpdf['D28'].astype(int).pct_change()*100),1)
-    kpdf['D36'] = round((kdata['report.overall'].astype(int).pct_change()*100),1)
+    kpdf['D30'] = round((kpdf['D26'].astype(float).pct_change()*100),1)
+    kpdf['D31'] = round((kpdf['D27'].astype(float).pct_change()*100),1)
+    kpdf['D32'] = round((kpdf['D28'].astype(float).pct_change()*100),1)
+    kpdf['D36'] = round((kdata['report.overall'].astype(float).pct_change()*100),1)
     kpdf['D38'] = round(((kdata['report.overall'].astype(int))/(kdata['report.turnover_total'].astype(int))),1)
     kpdf['D39'] = round(((kdata['report.grants'].astype(int))/(kdata['report.turnover_total'].astype(int))*100),1)
     kpdf['D40'] = round(((kdata['report.turnover_total'].astype(int))/(kdata['profile.sum_eme_kispe'].astype(int))),1)
